Server/eb-flask/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==2.0.3
Flask-Cors==4.0.0
fsspec==2024.2.0
gunicorn
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai
openai-whisper==20231117
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
sympy==1.12
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==2.2.2
